+++ src/clevertap.js
@@ -38,12 +38,14 @@ import {
   GCOOKIE_NAME,
   QUALIFIED_CAMPAIGNS,
   BLOCK_REQUEST_COOKIE,
-  ISOLATE_COOKIE
+  ISOLATE_COOKIE,
+  ENABLE_TV_CONTROLS,
+  WZRK_GEO
 } from './util/constants'
 import { EMBED_ERROR } from './util/messages'
 import { StorageManager, $ct, getMuteExpiry, isMuted } from './util/storage'
 import { addToURL, getDomain, getURLParams } from './util/url'
-import { getCampaignObjForLc, setEnum, handleEmailSubscription, closeIframe } from './util/clevertap'
+import { getCampaignObjForLc, setEnum, handleEmailSubscription, closeIframe, dismissActiveCampaigns } from './util/clevertap'
 import { compressData } from './util/encoder'
 import Privacy from './modules/privacy'
 import NotificationHandler from './modules/notification'
@@ -56,6 +58,7 @@ import encryption from './modules/security/Encryption'
 import { checkCustomHtmlNativeDisplayPreview } from './util/campaignRender/nativeDisplay'
 import { checkWebPopupPreview } from './util/campaignRender/webPopup'
 import { reconstructNestedObject, validateCustomCleverTapID } from './util/helpers'
+import TVNavigation from './modules/tvNavigation'
 
 export default class CleverTap {
   #logger
@@ -72,6 +75,7 @@ export default class CleverTap {
   #dismissSpamControl
   enablePersonalization
   #pageChangeTimeoutId
+  #tvNavigation
   #enableFetchApi
   #enableEncryptionInTransit
   #domainSpecification
@@ -146,7 +150,6 @@ export default class CleverTap {
     // Custom Guid will be set here
 
     const result = validateCustomCleverTapID(clevertap?.config?.customId)
-
     if (!result.isValid && clevertap?.config?.customId) {
       this.#logger.error(result.error)
     }
@@ -174,6 +177,7 @@ export default class CleverTap {
       session: this.#session,
       isPersonalisationActive: this._isPersonalisationActive
     })
+    this.#tvNavigation = new TVNavigation(this.#logger)
     this.enablePersonalization = clevertap.enablePersonalization || false
     this.event = new EventHandler({
       logger: this.#logger,
@@ -226,6 +230,11 @@ export default class CleverTap {
     this.spa = clevertap.spa
     this.dismissSpamControl = clevertap.dismissSpamControl ?? true
 
+    if (clevertap.config?.enableTVControls) {
+      StorageManager.saveToLSorCookie(ENABLE_TV_CONTROLS, true)
+    } else {
+      StorageManager.saveToLSorCookie(ENABLE_TV_CONTROLS, false)
+    }
     this.user = new User({
       isPersonalisationActive: this._isPersonalisationActive
     })
@@ -590,6 +599,14 @@ export default class CleverTap {
         this.#sendLocationData({ Latitude: lat, Longitude: lng })
       } else {
         if (navigator.geolocation) {
+          const isSafari = /^((?!chrome|android).)*safari/i.test(navigator.userAgent)
+          if (isSafari) {
+            try {
+              if (localStorage.getItem(WZRK_GEO) !== null) {
+                return
+              }
+            } catch (e) {}
+          }
           navigator.geolocation.getCurrentPosition(showPosition.bind(this), showError)
         } else {
           console.log('Geolocation is not supported by this browser.')
@@ -600,6 +617,7 @@ export default class CleverTap {
     function showPosition (position) {
       var lat = position.coords.latitude
       var lng = position.coords.longitude
+      try { localStorage.setItem(WZRK_GEO, 'true') } catch (e) {}
       $ct.location = { Latitude: lat, Longitude: lng }
       this.#sendLocationData({ Latitude: lat, Longitude: lng })
     }
@@ -608,6 +626,7 @@ export default class CleverTap {
       switch (error.code) {
         case error.PERMISSION_DENIED:
           console.log('User denied the request for Geolocation.')
+          try { localStorage.setItem(WZRK_GEO, 'false') } catch (e) {}
           break
         case error.POSITION_UNAVAILABLE:
           console.log('Location information is unavailable.')
@@ -734,7 +753,13 @@ export default class CleverTap {
     }
   }
 
-  init (accountId, region, targetDomain, token, config = { antiFlicker: {}, customId: null, isolateSubdomain: false, domainSpecification: null }) {
+  init (accountId, region, targetDomain, token, config = {
+    antiFlicker: {},
+    customId: null,
+    isolateSubdomain: false,
+    enableTVControls: false,
+    domainSpecification: null
+  }) {
     if (config?.domainSpecification) {
       this.domainSpecification = config.domainSpecification
       this.#session.domainSpecification = config.domainSpecification
@@ -759,6 +784,15 @@ export default class CleverTap {
 
     if (accountId) {
       encryption.key = accountId
+    }
+
+    const enableControls = StorageManager.readFromLSorCookie(ENABLE_TV_CONTROLS) ?? false
+    if ((config?.enableTVControls) || enableControls) {
+      // CleverTap handles navigation
+      StorageManager.saveToLSorCookie(ENABLE_TV_CONTROLS, true)
+      this.#logger.debug('CleverTap TV Navigation Mode: CleverTap will handle all navigation')
+      // Initialize CleverTap TV navigation system
+      this.#tvNavigation.init()
     }
 
     StorageManager.removeCookie('WZRK_P', window.location.hostname)
@@ -908,6 +942,13 @@ export default class CleverTap {
 
   pageChanged () {
     const currLocation = window.location.href
+    if (
+      this.#isSpa &&
+      this.#previousUrl !== undefined &&
+      this.#previousUrl !== currLocation
+    ) {
+      dismissActiveCampaigns()
+    }
     const urlParams = getURLParams(currLocation.toLowerCase())
     // -- update page count
     const obj = this.#session.getSessionCookieObject()